Make `Undef` optional.
The epp spec says:
```
 *  A <value> element that identifies a client-provided element
    (including XML tag and value) that caused a server error
    condition.
```

The current implementation assumes there will always be a `<undef/>`
element in the `<value>` element.

This causes deserialization to fail when the response includes tags and
values referencing the data that caused the error, instead of the
`<undef/>` element.

By making `Undef` optional, we can just skip deserializing any child
elements of `<value>`.